update WIT app tools files and name
